ai-log-analysis/requirements.txt
pyyaml==6.0.1
pytest==7.4.3
pytest-cov==7.0.0
requests==2.31.0
tzdata==2025.3
